Stop pruning vclocks on read repair.
This fixes a case of endless read-repair.  If the get is on
an object with a vector clock that will be truncated on the
next put (e.g. after being allowed to grow temporaruily large
with the young vclock logic) and one of the vnodes holds
an ancestor value then the read repair to the out of date
node will be truncated and in conflict with the original
value again.  Each get of the object will trigger a read
repair.

It is safer to allow the vclocks to grow a little more
in the read repair case than to truncate.