reduce function to be used in computing key counts
The intention of this patch is to provide a suitable function for
counting the keys in a bucket via keylist+map/reduce.  Using this
function saves bandwidth over counting on the client side (with
?keys=stream, for example) because keys are never sent to the client.
Using this function saves I/O over a query including a map phase because
the values stored under the keys are never read from disk.

A few ways to use this function:

Riak console:

    {ok, [KeyCount]} =
        C:mapred(<<"my_bucket">>,
                 [riak_kv_mapreduce:reduce_count_inputs(true)]).

HTTP (response body is a 1-element JSON list):

    $ curl -X POST -H "content-type:application/json" \
      http://localhost:8098/mapred --data @-
    {"inputs":"my_bucket",
     "query":[{"reduce":{"language":"erlang",
                         "module":"riak_kv_mapreduce",
                         "function":"reduce_count_inputs",
                         "keep":true}}]}

HTTP with key filters:

    $ curl -X POST -H "content-type:application/json" \
      http://localhost:8098/mapred --data @-
    {"inputs":{"bucket":"my_bucket",
               "key_filters":[["starts_with","f"]]},
     "query":[{"reduce":{"language":"erlang",
                         "module":"riak_kv_mapreduce",
                         "function":"reduce_count_inputs",
                         "keep":true}}]}