Bug 808 - riak_kv_vnode:do_list_keys/6 and bucket-aware optimizations
Add an iterative process to try the three different API variations for
folding over keys in various backends.  IIRC there's a different
bug/change-request to change these APIs to be more harmonious, so we'll
tackle that bridge when we take it by the horns.

Also added more extensive backend unit tests: for bitcask, cache, dets,
ets, fs, and gb_trees backends.  And added new-style folding API to the
ETS backend, as an example (since it doesn't do anything differently than
the fold implementation).