Let riak_client:get optionally return {r_value_unsatisfied, R, P}, where P is the number of ok replies, and R is the requested R value.
This changes the riak_kv_get_fsm logic in the cases where it receives 
{error,  _} to 

 - not reply {error, notfound} unless the error is the Nth reply
 - reply {r_value_unsatisfied, R, P} iff 0 < P < R and all errors
   from vnodes were notfound's, i.e. it is likely that a read-repair
   is imminent which will fix the issue.

When the webmachine front receives a {r_value_unsatisfied, R, P},
it will issue an {halt, 503}.

So, this slows down the failure cases somewhat: it is slower to see the difference between a 404 "really-nothing-there" and a 503 "we have something, but the ?r=R request parameter was not satisfied.

Four our usage cases, we really need to be able to make this distinction.