Last refactor broke deletes - fixed.
Deletes were being issued synchronously, one vnode at a time.
The result was never being checked and nothing blocks another
get from trying to delete them too.  They may as well be async
so they execute closer together.  This will be further hardened
when we add an if-match on the vclock.