link mapper to phase pid and JS VM
Linking to the phase pid ensures that this mapper gets torn down
immediately when the phase dies, instead of hanging around finishing its
current evaluation and batch.

Linking to the JS VM is the easiest way to "return" the VM to the pool.
It doesn't actually return the VM, but instead kills it and allows the
VM supervisor to spin up a fresh one.