Added {error, {r_val_unsatisfied, R, Responses}} as proposed by Kresten.
The implementation is slightly different and takes account of
deleted objects - they count towards notfounds.  That was the only
way that made sense in the rapid-cycling delete case where
some nodes may have removed the object and others have it present.

The {error, [VnodeErrs]} return has vanished for now.  The information
will be available with the new details option when the return becomes
a 3-tuple (probably containing [{vnodes, [ok,notfound,Error]}].