Unique list of siblings while merging.
Prior to the fix for bz://977 when performing syntactic put merge
each of the vndoes used a local timestamp for the vclock increment.
If the vnodes were not updating inside the same second the siblings
would be duplicated.   This will repair any objects with duplicated
siblings.

Thanks to Kresten for suggestions.