Add proper start_link function to riak_kv_delete
Sometime in February we put a simple_one_for_one supervisor in place
to spawn riak_kv_delete instances.  The problem is that the child spec
specifies a start fun which also does the work.  This means that the
work will be done on the supervisor container's process which isn't
good and it breaks the supervisor interface (i.e., the start function
should return {ok,Pid} and do as little work as possible).

This actually causes delete to fail when the Search KV hook is
installed because there is a nested delete inside the commit hook that
is blocked because the delete supervisor's process is waiting for the
first delete request to finish processing.