Add Reply parameter to handle_not_found_reply so that the notfound errors are present if requested by the client.
tags: az327

The fix to address bz1006 causes notfound error information to be excluded
from the results of a MapReduce even if the inclusion is requested by
the client. This changes corrects that by adding a Reply parameter to
the handle_not_found_reply function and ensuring that after the
final preference list entry has been checked that Reply is appended
to the phase state's pending value.

Add test cases for riak_kv_mapreduce:map_object_value to show that
the include_notfound and filter_notfound actions work as expected.