Change from client-based to vnode-based vclocks.
Changes put to work as described in the Dynamo paper.  Put requests are
coordinated by a node in the preference list.  The coordinator writes to
the local vnode first and then sends to the remote vnodes.  By writing
to the local vnode first and assigning vnodes a vclockid the first
time data is stored in the partition, the vclock can be incremented
in the vnode to guarantee a frontier vclock so the object will be
stored on another node.

This removes the requirement for clients to be 'well behaved' and always
get before they put.  If allow_mult is true then a sibling will be
created for each put.  It also resolves some edge cases for eventual
consistency properties where clients could see both sides of a partitioned
cluster.