Improve error handling and process interaction in the MapReduce code to prevent the termination of all active MapReduce jobs if a single MapReduce job times out or terminates unexpectedly.
Fixes: AZ340

Handle the case of an error tuple being returned from the calls to
riak_kv_mapper_sup:new_mapper in riak_kv_map_master so that a timeout
does not cause riak_kv_map_master to terminate and take all active
MapReduce jobs with it.

Change riak_kv_map_master so that it does not monitor mapper processes
and instead trap exits and rely on the link to the mapper proesses
to detect when the mapper process terminates.