First stage refactor of get core logic.
Preparing to move get logic out to separate module for reuse
by other portions of the system.

Found an edge case bug with deletes in the process.  Deletes were
being issued when the merged object was deleted and all vnodes were
primaries.  Deletes should only be issued once all vnodes have
a tombstone with the same vclock.  Delivery of delete messages
is not guaranteed so if a vnode was storing stale data and the
delete message was not executed (for example, node out of disk space)
then the object would be resurrected next access.