Got handoff working with object going immediately undefined.
In the real code the handoff transmission/deletion is async.
Handoff starts, network connection gets set up, a fold kicks off
and once all data is transmitted the vnode does the drop.  It could
service other requests in the mean time.  In fact it looks
like the data is dropped on handoff completion, even if something
has arrived in the mean time.