Erase the reference to a map executor in the case of a notfound instead of decrementing the input counter.
Fixes: az445 bz1112

When a notfound error occurs during a map phase the phase creates a new map plan and retries using a difference preference list entry. A lingering reference to the original map phase executor used to track input counts was the source of the error. The solution is to remove the reference to the original map phase executor in the case of a notfound error instead of simply decrementing the associated counter.