Catch exit with reason exhausted_preflist on the call to schedule_input after a nodedown message is received.
Fixes: az449 bz1113 bz1114

The schedule_input function in riak_kv_map_phase calls itself
in the case that it catches a nodedown exit while trying to
start mapper processes. If all the primary preference list
entries have been checked or their nodes are down this inner
call to schedule_input will result in an exit with reason
exhausted_preflist. Instead of just having the map phase
process exit, this change catches the exit and returns an
error to the caller of schedule_input so that notfound results
can be returned in the case of bz1113 or a more descriptive
error message can be return in the case of bz1114.

Do not expect an empty list for the ClaimList
parameter in the termination clause of
riak_kv_mapred_planner:claim_keys.

Add extra function clause for riak_kv_mapred_planner:claim_keys.