AZ 469: function inputs in pipe mr emulation
Add support for `{modfun, Mod, Fun, Arg}` tuple as input to a MapReduce
job (rather than bucket name or list of BKeys).

Examples of this use are sparse.  Further compatibility tests will likely
be tricky because the current implementation is very closely tied to the
`luke` app.  (Similarly, this implementation will be closely tied to the
`riak_pipe` app because the generator function is expected to call
`riak_pipe:queue_work()` and `riak_pipe:eoi()` directly.)