Backend API standardization and refactoring - Phase 1 and Phase 2
Fixes: AZ543 AZ596

This change set includes the following changes:

* Updates all backends to use a standard api
* Removes the gb_trees, cache, fs, ets, and dets backends
* Consolidates cache and ets backend functionality into a memory backend
* Updates riak_kv_vnode to use the new backend api
* Change bitcask backend to use a symlink to point to the active directory to facilitate drops.
* Add a module for buffered return of backend fold results.
* Updates to relevant eqc and eunit tests.
* Specs for the backend api functions in all backend modules.

This change set also adds support for secondary indexing to the
backend api and makes changes to riak_kv_vnode and the secondary index
code to use these changes. These changes include a number of
optimizations to improve the performance of the index backend
including discontinuing the use of a proxy object to store the index
information in merge_index.