Add `disable_hooks` option for put FSM
In some cases you may want to disable the precommit/postcommit hooks.
For example, when using cluster replication you don't want an
incoming, replicated object to trigger hooks again since repl
uses a hook and this would cause every replication request to
prepetuate.

One could be more sophisticated by allowing to pass in a black/white
list of disable/enabled hooks but for the given use case this is
not needed and I felt it's better not to complicate things if it's
not merited (YAGNI).