Backend Refactor - Phase 3
Fixes: az603

Adds support for asynchronous folds for the backends. The async
behavior is used by default if it is available in the backend (this is
the case for bitcask, eleveldb, and memory), but can be disabled using
an entry in the riak_kv section of app.config. Setting async_folds to
false will override the default.

This pull request also updates the backend_eqc module to test the
async folds and bitcask, eleveldb, and memory backends now run
quickcheck tests using both sync and async folds.