Updated ec_eqc and put fsm/vnode to add vnode epoch to vclock node id.
Fixes the case identified by Bryan in the pull request comments, where
a vnode can transfer ownership (deleting the local copy of the data) then
be used as a partitioned fallback.  Previously that would have re-used
a vclock entry, but now the vnode epoch is included in the clock
to guarantee uniqueness.