don't block PB socket while sending pipe mapred inputs
If the PB socket-handling process is blocked sending pipe inputs, it
can't tell if the client disconnects before all of the inputs are sent.
This can lead to a pipeline living long after the thing waiting for it
has vanished.  The patch sends inputs from another process instead, and
kills that other process if the client disconnects before all inputs
are sent.