link async input sender to pipe builder
This way, if the pipe shutsdown for any reason, the sender is
automatically cleaned up.  This also removes the need to remember to
kill the sender from riak_kv_pb_socket and riak_kv_wm_mapred, because
their death will take the builder down, which will also now take the
sender down.