enable error traces to catch JS eval timeouts in pipe mapred
This patch catches timeout errors in Javascript evaluation, and sends
an error trace when they happen.  The patch also turns on captures of
error traces at the sink so these timeouts (and other errors) can be
reported to clients.

This issue was found by the Java unit test
ITestMapReduce.timeoutMapReduceThrowsTimeoutException.