Remove the use of symlinks for data directories in the bitcask backend.
Fixes: az677 bz1177

Remove the symbolic links from the partition name to the actual active data
directory for the bitcask backend. The original intent was to facilitate
drop calls with bitcask and avoid file handle problems, but the links
are no longer used.

Also improve checking for existing data directories to use. This includes
a check for a directory with just the bare partition name which facilitate
rolling upgrades. Each new data directory is the name of the partition with a
timestamp appended. When the bitcask backend is started if no directory exists
with the bare partition name, the directory with the most recent time stamp
appended is opened and used as the current data directory. A log message
listing any unused directories for the partition is also generated.