Add a coverage fsm and plumbing for getting backend status from all vnodes.
Fixes: az722

This commit add riak_kv_backend_status_fsm and makes other necessary
changes to provide a way to query the backend status for all available
vnodes.  There is also the addition of a backend_status function to
riak_kv_console so this information can be made available via a
command line tool such as riak-admin.