Fix problem with repeated bitcask merges.
Fixes: bz1200 az728

The bitcask backend is not passing the correct path to the bitcask
merge worker and as a result the hint and data files are created in
the wrong place. Bitcask then repeatedly tries to merge even though
from the end user perspective it should not need to.  The fix is to
simply pass the correct path to merge worker.