Add a function to riak_kv_console to display the status of local vnodes.
Fixes: az722

This change adds a vnode_status function to riak_kv_console. It also
add code to riak_kv_vnode to handle a request for vnode status
information. Finally a vnode_status function is added to
riak_kv_status to determine the local vnodes, send status requests to
each one and accumulate the results. Currently the only piece of vnode
status information that is available is the backend status.