Fix for bucket listing in eleveldb.
Fixes: az731 bz1202

The exception handling to catch when eleveldb exits a fold early was
omitted from the fold_buckets function, but with the addition of the
secondary index capability to the backend, bucket listing can break out
of fold_buckets early. This change adds that exception handling to
the fold_buckets function.