Relocate wrapping of handoff fold functions in riak_kv_vnode.
Fixes: az736 bz1214

The function used for folding objects during handoff or replication
needs to be wrapped since the riak_kv backends expect a fold function
of arity 4, but riak_core presents a function of arity 3. This wrapping
was formerly handled in the handle_handoff_command and fold functions, but
this can cause problems when handoff and fullsync replication are occurring
simultaneously. This change moves that function wrapping the handle_command
clause that handles object fold requests in the case of both handoff
and replication and alleviates the problems.