prevent connection reuse after pipe mapred error (BZ1222)
Because of the way pipe sends results, sending the response to the client
before eoi means that the request process may still have messages in its
mailbox.  If mochiweb tries to keep the connection alive, its next loop will
receive one of these messages instead of HTTP/TCP data, which will cause it
to send a 400 to the client.  The interim fix here is to just tear down the
connection, unconditionally, after the initial mapred error is sent, forcing
the client to reconnect for its next request.