Add delete_mode to control when backend deletion takes effect.
With the vnode-based vclock scheme all update/delete cycles against
a single object will use the vnode ids.  If one of the vnodes is
running slow there is a small window where the tombstone has been
removed from some nodes but not others.  If an object is created
on a vnode where the object has been removed then the vclock entry
for that vnode will be less than the tombstone and the write would
be ignored.

This patch adds a delete_mode app env that controls when the tombstone
object is removed from the backend.  If 'keep' then tombstones
are always kept, which is the safest option, but an object will always
consume disk space once created.  'immediate' mirrors the 0.14.2
behavior of removing the tombstone once the vnode delete request
is received.  The middle ground is to set a delay time which
makes the vnode wait that long and only remove the object if it
has not been modified in that time.  Setting to 1000ms was sufficient
to pass the bz://260 delete_bug script.

If a delay is used, the vnode uses erlang:send_after to trigger the
deletion.  For a system that is under very high delete load
there could be many timers created, however they should be fairly
cheap.