Pass [asis] option to put FSM if riak_client has clientid.
In a mixed cluster, if a 1.0.0 node is coordinating a put that it
is not in the preference list for and decides to forward to a pre-1.0.0
node it will crash as riak_kv_put_fsm_sup does not exist on the remote
node.

With the asis option the put FSM will not attempt to forward and
write to the vnodes in the preference list directly.

Fixes: bz://1234