BZ1254: Expose tombstones in siblings and link-walking queries over HTTP.
As discussed in http://issues.basho.com/1254, tombstones that have not
been reaped may appear in link-walking queries but only appear to be
empty objects with a Content-Type of "application/octet-stream". This
adds the X-Riak-Deleted header to multipart bodies emitted by
link-walking queries (and consequentially siblings that are tombstones
in normal fetches).