Make riak_kv register webmachine routes after kv_sup started.
The riak_kv webmachine resources require parts of the riak_kv
application to be started (riak_kv_*_fsm_sup).  Before the change,
a node starting under load would fail requests until the supervisors
were started.

Now the webmachine routes are registered after the sup has started.
This has the side effect that routes will not be re-registered
when the riak_kv_sup is restarted, however there is no code
to remove the routes from webmachine when the application stops
so the change should not have any unintended consequences.