Clean up data dir before running tests
The tests were failing because a the ring used by a previous test was being
picked up after the riak_kv application is started.  If you look in the
keys_fsm_eqc you can see it performs a ring_trans to assign all partitions to
new node.  I opted to just wipe the data dir from the previous run entirely.