Set `small_vclock` equal to `big_vclock`
The default value of `small_vclock` has been changed to be equal to
`big_vclock` in order to delay or even prevent unnecessary sibling
creation in a Riak deployment with bidirectional cluster replication.
When you replicate a pruned vector clock the other cluster will think
it isn't a descendent, even though it is, and create a sibling.  By
raising `small_vclock` to match `big_vclock` we reduce the frequency
of pruning and thus siblings.  Combined with vnode vclocks, sibling
creation, for this particular reason, may be entirely avoided since
the number of entries will almost always stay below the threshold in a
well behaved cluster (i.e. one not under constant node membership
change or network partitions).

See also: https://issues.basho.com/show_bug.cgi?id=1310

Note that this is not a fix for the above issue, just an attempt to
reduce it's likelihood.