make JS VMs monitor the processes that check them out
This is to prevent VMs from leaking out of the pool when their owner
dies before triggering their checkin.  For example, if a map process
dies after requesting a vm reservation, but before dispatching any js
evaluations.  There is also the case where a dead process (because the
pipe was stripped down before the reserve call was fulfilled) can
reserve a VM.  This patch also prevents that from happening since the
VM will receive a DOWN msg and return itself to the pool