Use a call instead of cast to checkout VM
If you use a cast to checkout the VM there is a chance that two or
more processes could checkout the same VM.  This is because the
process that sends the `checkout_to` msg and the process that sends
the `blocking_dispatch` method are different and therefore the msgs are
not guaranteed to be ordered in relation to one another.  I.e. a race
can occur between the two messages.

1. worker A calls `reserve_vm`

1a. pool mgr removes VM from pool and sends async `checkout_to`

2. worker A calls `blocking_distpatch` (`checkout_to` still in flight)

3. VM does work on behalf of A, puts itself back in the pool, and
sends reply to worker (`checkout_to` still in flight)

4. worker B calls `reserve_vm` and gets same VM that worker A used
(`checkout_to` from original checkout still in air)

4a. pool mgr removes VM from pool and sends async `checkout_to`

5. original `checkout_to` lands on VM, worker A is dead

6. VM gets DOWN msg because worker A is dead

7. VM checks itself back into pool (but worker B thinks it has sole
possession)

8. worker C calls `reserve_vm` and gets same VM as worker B

Now two workers have the same VM.  This interleaving is probably rare
and probably wouldn't cause damage but it seems better to just use a
call instead to checkout the VM so that you know if a VM is checked
out it knows who it's owner is.  Since an idle VM shouldn't be
blocking on any other work the `checkout_to` call should return
promptly and shouldn't cause a long wait in `reserve_vm`.  To be safe,
a timeout is used.