Fix predicate error in remove_indexes/3.
Before this change, the model state was having all indexes
corresponding to the deleted key (in any bucket) or bucket (with any
key) removed. This meant that when index specs were being calculated
for removal in delete/5, sometimes none were there to remove,
resulting in the backend not being empty after all keys were deleted.