Add repair ability
Add the ability to repair a KV partition.  If a partition's data is
lost or corrupted this command may be used to rebuild it from adjacent
partitions.  This is preferable over a list-keys + read-repair since
it is more targeted, should complete more quickly and should put less
strain on the cluster.