catch noproc in index and listkeys coverage pipe link (basho/riak_pipe#48)
It is possible that the coverage fsm handling the index query at the head of a pipe
will finish (or otherwise exit) before we try to link to it. This patch wraps the
link in a try-catch to prevent a 'noproc' exit (if the coverage fsm did finish
successfully faster than expected, we want to pick up the 'done' message from our
inbox).

The listkeys pipe coverage fitting had the same potential problem.