separate the socket handler and sink processes in MR endpoints
This is relevant to riak_kv#366's interests.

Prior to this commit, the process holding the HTTP/PB socket was also
the process acting as the Pipe sink for MapReduce requests. This led to
many problems, including overwhelming mailboxes with error logging, and
extra unexpected messages that had to be ignored if the socket was kept
open to handle other requests.

This commit creates a new process to act as the sink, and sets up a
request-response interface between them. The sink process spools pipe
output until the socket-handling process asks for them.

This removes our dependence on the undocumented
mochiweb_request_force_close feature, which was used to forcibly close
an HTTP connection after a MapReduce failure to avoid incorrect 400
errors later.