check the result of queue_work while sending mapred inputs
Without this, a process might continue to send inputs to a MapReduce
pipe for a long time after the pipe has disappeared.

This change should clear up the majority of cases where a log is spammed
with messages about "fitting was gone before startup". There may still
be N (one from each of the N primaries of a preflist) as the call must
fail once before triggering the exit, but it will not continue through a
large input list.