Change riak_kv_bitcask_backend to use bitcask:is_empty_estimate
Previously, determining if a bitcask was empty or not was accomplished
through a fold over the keydir which is a potentially blocking operation.
This commit changes riak_kv_bitcask_backend:is_empty to use the new
function bitcask:is_empty_estimate.

The estimate is determined from the bitcask stats, which may overcount
data, but will not undercount. Therefore, the estimated result may return
false when the bitcask is actually empty, but it will never return true
when there is data. In all cases where is_empty is currently used in
Riak, an estimate is acceptable. In the worst case, additional work may
be triggered that is unnecessary but safe (eg. folding over an empty
bitcask).

See issue: basho/riak_kv#423