Ensure that index hashtree failures do not crash vnodes
Previously, riak_kv_vnode would create riak_kv_index_hashtree PIDs
using spawn_link, which would force the vnode to crash if the hashtree
ever crashed. Now, the vnode spawns and monitors the hashtree PID and
will recreate the tree if it ever dies. If creating a new tree fails,
the vnode will retry to create the tree periodically. This handles the
case where a lingering LevelDB lock has yet to be released between
when a previous tree exited and the new tree was created. The first
create may fail, but eventually the lock is released and a new tree
started.