Make AAE exchange check riak_kv service status
Exchanges will now only occur if both the local and remote nodes are
up and have the riak_kv service running. This prevents AAE from trying
to do exchanges with nodes that are offline, or performing exchanges
with a node that is up but not handling K/V requests and therefore
would not repair anything even if read-repair was triggered.