Add basic health check to Riak KV
This commit adds a simple health check to riak_kv to support the new
health check functionality added to riak_core. This health check
monitors the process mailboxes of k/v vnodes. The check fails if any
vnode on the node has a mailbox containing over a configurable limit
of messages -- default 5000.