Add AAE status subsystem and related console command
Add riak_kv_entropy_info module that provides an API to keep track of AAE
status such as tree build times and exchange statistics. The information
is stored in a public ETS table that is owned by the riak_kv_sup.

Update the AAE subsystem to call into the info module where appropriate.

Add aae_status/1 as a console command in riak_kv_console that pretty prints
the AAE information.