Revert to using quorum reads for AAE read repair
The commit that fixed AAE from repairing non-divergent keys over and
over also changed the AAE triggered reads to use R=all. The intention
was to ensure that any potential read repair was triggered before
triggering an update of the respective AAE trees.  However, this is
not necessary. If a rehash happens before the read repair, the newer
repaired hash will be written during the vnode put. This commit
reverts back to using quorum reads, reducing the impact of slow
replicas on AAE repair.